% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacetimetabs.R
\name{names<-.SpacetimeviewList}
\alias{names<-.SpacetimeviewList}
\title{Set tab titles for a SpacetimeviewList using names}
\usage{
\method{names}{SpacetimeviewList}(x) <- value
}
\arguments{
\item{x}{A SpacetimeviewList object}

\item{value}{Character vector of tab titles}
}
\value{
The SpacetimeviewList with updated tab titles
}
\description{
Set tab titles for a SpacetimeviewList using names
}
