% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacetimetabs.R
\name{plot.SpacetimeviewList}
\alias{plot.SpacetimeviewList}
\title{Plot method for SpacetimeviewList}
\usage{
\method{plot}{SpacetimeviewList}(x, ...)
}
\arguments{
\item{x}{A SpacetimeviewList object}

\item{...}{Additional arguments passed to print}
}
\value{
The SpacetimeviewList object (invisibly)
}
\description{
Converts a SpacetimeviewList to a spacetimetabs widget and displays it
}
