% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacetimetabs.R
\name{save.SpacetimeviewList}
\alias{save.SpacetimeviewList}
\title{Save method for SpacetimeviewList}
\usage{
save.SpacetimeviewList(x, file, ...)
}
\arguments{
\item{x}{A SpacetimeviewList object}

\item{file}{The file path to save the HTML file to}

\item{...}{Additional arguments passed to print}
}
\value{
The SpacetimeviewList object (invisibly)
}
\description{
Converts a SpacetimeviewList to a spacetimetabs widget and displays it
}
