% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacetimetabs.R
\name{spacetimetabs}
\alias{spacetimetabs}
\title{SpaceTimeTabs}
\usage{
spacetimetabs(
  ...,
  tab_titles = NULL,
  width = "100vw",
  height = "100vh",
  elementId = NULL
)
}
\arguments{
\item{...}{spacetimeview objects to combine}

\item{tab_titles}{Character vector of tab titles. If not provided, default titles will be used.}

\item{width}{Width of the widget}

\item{height}{Height of the widget}

\item{elementId}{Optional element ID for the widget}
}
\description{
Combine spacetimeview objects to create a page with tabs.
}
